requirements.txt
Flask==3.0.0
pysaml2==7.4.2
webauthn==2.1.0
pymongo==4.6.1
python-dotenv==1.0.0
cryptography==41.0.7
